+++ UMeR/Viagem.java
@@ -13,12 +13,12 @@ public class Viagem
     private Coords fim;
     private int distancia;
     private int precoAcordado;
-    private int precoFinal;
+    private float precoFinal;
     private Motorista condutor;
     private Veiculo veiculo;
     private Date data;
 
-    public Viagem(int id, Cliente cliente, Coords inicio, Coords fim, int distancia, int precoAcordado, int precoFinal, Motorista condutor, Veiculo veiculo, Date data) {
+    public Viagem(int id, Cliente cliente, Coords inicio, Coords fim, int distancia, int precoAcordado, float precoFinal, Motorista condutor, Veiculo veiculo, Date data) {
         this.id = id;
         this.cliente = cliente;
         this.inicio = inicio;
@@ -43,7 +43,7 @@ public class Viagem
         return fim;
     }
 
-    public int getPrecoFinal() {
+    public float getPrecoFinal() {
         return precoFinal;
     }
